install: fix octoprint start detection
Octoprint startup monitoring currently rely on a specific log line being
detected. The configured log line is deprecated, current versions of
octoprint have different logs.

Update the monitored log line to allow the installer to succeed. This is
only a hotfix: the startup detection _must_ be reworked to rather use a
proper API to detect octoprint startup, as logs are not a stable API.